requirements.txt
python-fasthtml
uvicorn
google-genai>=1.0.0
pydantic-ai
pydantic>=2.0.0
fpdf2>=2.7.0
qrcode[pil]>=7.0.0
pychord>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0duckduckgo-search
